restore missing code

+++ violet-framework/src/main/java/com/horstmann/violet/framework/file/GraphFile.java
@@ -45,10 +45,6 @@ public class GraphFile implements IGraphFile
         try
         {
 			this.graph = graphClass.newInstance();
-			this.autoSaveFilename = new SimpleDateFormat("yyyyMMdd_HHmmss").format(Calendar.getInstance().getTime()) + ".html";
-
-			this.autoSaveFile = new File(this.autoSaveDirectory + this.autoSaveFilename);
-			this.autoSaveFile.createNewFile();
         }
         catch (Exception e)
         {
@@ -75,10 +71,7 @@ public class GraphFile implements IGraphFile
         if (in != null)
         {
 			this.graph = this.filePersistenceService.read(in);
-			this.autoSaveFilename = file.getFilename();
-
-			this.autoSaveFile = new File(this.autoSaveDirectory + this.autoSaveFilename);
-			this.autoSaveFile.createNewFile();
+	
         }
         else
         {
@@ -86,6 +79,7 @@ public class GraphFile implements IGraphFile
                     fileOpener.getFileDefinition().getDirectory());
         }
     }
+
 
     @Override
     public IGraph getGraph()
@@ -157,8 +151,18 @@ public class GraphFile implements IGraphFile
     }
     
 	@Override
-	public void autoSave() {
+	public void autoSave(String fileDirectory) {
 		try {
+		
+			if (this.autoSaveFileName == null) {
+					this.autoSaveFileName = (this.currentFilename == null ? java.util.UUID.randomUUID().toString() : this.currentFilename) + new SimpleDateFormat("yyyyMMdd_HHmmss").format(Calendar.getInstance().getTime()) + ".html";
+				}
+			
+				if ((autoSaveFile == null) || ((autoSaveFile != null) && (!autoSaveFile.getAbsolutePath().contains(fileDirectory) == false))) {
+				 	this.autoSaveFile = new File(fileDirectory + File.separator + this.autoSaveFileName);
+				 	autoSaveFile.createNewFile();
+				}
+			
 			if (autoSaveFile.exists()) {
 				JFileWriter jfilewriter = new JFileWriter(autoSaveFile);
 				this.filePersistenceService.write(this.graph, jfilewriter.getOutputStream());
@@ -170,8 +174,9 @@ public class GraphFile implements IGraphFile
 	
 	@Override
 	public void removeBackup() {
-		if (autoSaveFile.exists())
+		if ((autoSaveFile != null) && (autoSaveFile.exists())) {
 			autoSaveFile.delete();
+		}
 	}
 
     @Override
@@ -341,13 +346,11 @@ public class GraphFile implements IGraphFile
      * Needed to identify the physical file used to save the graph
      */
     private String currentFilename;
-    private String autoSaveFilename;
 
     /**
      * Needed to identify the physical file used to save the graph
      */
     private String currentDirectory;
-    private final String autoSaveDirectory = System.getProperty("user.home") + File.separator + "VioletUML" + File.separator;
 
     private boolean isSaveRequired = false;
 
@@ -387,4 +390,6 @@ public class GraphFile implements IGraphFile
     private List<IGraphFileListener> listeners = new ArrayList<IGraphFileListener>();
 
     private File autoSaveFile;
+    private String autoSaveFileName;
+
 }
